requirements.txt
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
packaging==23.0
pydantic==1.10.7
python-dotenv==1.0.0
python-telegram-bot==20.2
replicate==0.7.0
requests==2.28.2
rfc3986==1.5.0
sniffio==1.3.0
typing_extensions==4.5.0
urllib3==1.26.15